refactor: Extract log_tick into new file - telemmetry_log.hpp
